initializing paypalSDK if wallet changes
previously we were checking whether the wallet is GENERAL and whether the channelId has changed. but channelId is getting changed even before the wallet is chenged to type GENERAL. so paypalSDK init was not happening. Now we are not only checking whether the channelId has changed, but also whether the wallet's type has been changed. if either is true, then we initialize the paypalSDK